realtek: Add support for LED configuration on RTL839x
This adds support for LED configuration on the RTL839x SoCs. As for the
RTL93xx and in contrast to the 838x, led configuration works via so-called
LED-sets.

The following is an example of a device node in the .dts for the
Netgear GS728TPv2 with 24 Ethernet ports displaying yellow for 10/100
MBit and green for 1000MBit. Additionally it has 4 SFP ports connected
to an RTL8214QF which also show green or yellow:

led_set: led_set@0 {
	compatible = "realtek,rtl8390-leds";
	leds_per_port = <2>;
	led_mode = "serial";
	led_set0 = <0x0a 0x0f 0x1f>;  // 10/100 (yellow) and 1000MBit (green)
};

The leds_per_port property defines the number of LEDs that are connected to each
port, in the given case 2 for green and yellow.
The led_mode property is one of "serial", "color-scan", "bi-color-scan".
The scan modes are implemented via an RTL8231 GPIO expander in the respective
mode, whereas the "serial" mode is implemented with a shift-register logic IC.

Each port uses a led-set (by default set 0), which can be configured
using the led_setX properties. On the RTL839x these are 3 values between 0
and 31. It is unknown how they work. They should be read out of the
RTL8390_LED_SET_x_y_CTRL registers in u-boot after issuing "rtk network on"
to enable the network and corresponding LED configuration in u-boot.

Signed-off-by: Birger Koblitz <git@birger-koblitz.de>