realtek: Fix DSA handling of SFP+ ports on RTL93xx
This fixes handling of the SFP+ ports in DSA by correcting
the output of the link-state routine. This is only needed when an
interface is either 1000BX or 10GBR, as other ports are always fixed
config and handled by the initial phy setup.

This probably needs to be done better in the future ...

Signed-off-by: Birger Koblitz <git@birger-koblitz.de>
Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>