realtek: Fix SFP+ port in Zyxel XGS1250-12
The SFP+ port is a managed port which does not have a fixed 10Gbit link
but can also hold 1000Base-X modules. Fix this in the .dts.

It's a bit odd to set the phy-mode at all, or set it to 1000base-x, as
10gbase-r technically is also right. However, currently, the 10g mode
causes a lot of spam in the console, and in the end, it doesn't matter,
as the phy-mode is set on hot-plug.

Signed-off-by: Birger Koblitz <git@birger-koblitz.de>
Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>