realtek: fix for ECS4125-10P setup
We copy the successful setup for SFP+ ports on the RTL9300. Also the hardware
configuration of this development device seems to have two additional SFPü
ports configured (which is also visisble from the casing), which however
are not equipped. Nevertheless they need to be taken into account to get
the HW configured (at least at the moment).

Signed-off-by: Birger Koblitz <git@birger-koblitz.de>