realtek: correct MIPS clocksource frequency on RTL931x
On the RTL931X, the frequency of the MIPS clock is only 1/4 of the CPU
frequency. Fix this when setting the frequency of the MIPS clock.

Note that the RTL931x also has a GIC timer as clocksource and for timer-
events. It's frequency is 1/2 the CPU frequency. For both
clocks event sources the rating is given as 200 + freq_in_kHz
giving for the MIPS clock 225, whereas the GIC rating is 250, which
make the GIC timer be selected as the scheduler clock.

The actual clock frequencies and especially the fact that the MIPS clock uses
only 1/4 of the CPU frequency was established by forcing a clock
source as scheduler clock and then comparing the duration of
e.g. sleep 10 with the actual length of 10 seconds.

The CPU clock frequency and the lexra clock frequency both come from
the output of u-boot.

Signed-off-by: Birger Koblitz <git@birger-koblitz.de>