realtek: add support for RTL838x port led configuration
This add support to configure the Port LEDs in such a way that they are automatically
controlled by the SoC on the RTL838x. The configuration routine combines information about
ports with configuration options describing the serial control via an RTL8231 or other kinds
of ICs with shift registers

To enable a node in the .dts is necessary, which could looks like this:
	led_set: led_set@0 {
		compatible = "realtek,rtl8380-leds";
		led_mode = "serial";
		combo_led_mode = <2>;
		led_set0 = <0x0a 0x0a>;
		led_set1 = <0x0f 0x0f>;
	};
The compatible node is necessary. The led mode can be one of serial, color-scan or bi-color-scan.
The combo_led_mode property is optional and defines the way combo leds (e.g. on an RTL8124FC) are
controlled. They can have either one set of LEDs or 2 sets: one above the SFP and one about the
copper ports. 3 leds are configurable which define the way certain conditions like 100MBit
are reflected by the LEDs. It is not clear how this is being encoded.
Additionally, every port can have a port property "led-num", which defaults to 2 and gives the
number of LEDs associated with a port, which can be 0-3. Default is 2 for backward compatibility.

The settings in the node need to come from the original software of the switch. The easiest is
if u-boot configure the LEDs correctly after an "rtk network on". In that case, read out
the LED configuration registers and fill in the values appropriately:
u-boot># md.l 0xbb00a000 4
bb00a000: 2f396b1b 00f501ea 0f00ffff 00000000    /9k.............
For example address bb00a0004 gives the configuration of the LED sets. Its value is in binary
binary values:			00000 01111 01010 00000 01111 01010
hex values of 5 bit fields: 	0x0   0xf   0xa   0x0   0xf   0xa
LED set 			2       1     0    2      1     0
port ranges:			24-27 24-27 24-27 0-23  0-23  0-23

If u-boot does not correctly configure the LEDs, then the information can be taken
out of the switch's configuration structure which can be found in the OEM image.

Signed-off-by: Birger Koblitz <git@birger-koblitz.de>