realtek: Add configuration support for the RTL8226
This adds support for the configuration of the RTL8226 PHYs, including
support for HiSGMII and SGMII connections used with the RTL9300 SerDes
for 2.5GBit and 1GBit modes. Setup includes TX configuration and RX
calibration of the PHY-MAC link. We provide for MDIO pin swapping as found in the
Zyxel XGS1210 switch.

Signed-off-by: Birger Koblitz <git@birger-koblitz.de>