realtek/rtl930x: Add shared base for xgs12xx series
The XGS1250-21, XGS1210-12 and XGS1010-12 are very similar, with very
minor differences. Lets have a shared devicetree to make maintaining
multiple devices with a common base easier.

The new shared devicetree also drops the compatible partition layout
with the OEM firmware, as it does not serve any purpose and only wastes
space.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>