realtek/rtl930x: Migrate XGS1250-12 to common devicetree
With the newly introduced shared devicetree for the XGS12xx series, also
switch the XGS1250-12 to use the shared devicetree.

While here, clean up a small ugliness, without changing actual
functionality.

Note, that due to the weird I2C muxing, there's some overlap there that
will be addressed in the future with a proper I2C driver that properly
handles this.

This change also includes a build to keep vendor U-Boot/flash layout
compatibility. The preferred method is to use the update U-Boot SPL.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>