realtek: add support for Zyxel XGS1010-12
The Zyxel XGS1010-12 is an 8x 1 GBit, 2x 2.5 GBit NBaseT ports and
2 SFP+ ports switch. It is identical to the Zyxel XGS1210-12, except for
the partition layout which is probably caused by the basic firmware
coming from the SDK with the default dual partition layout. To ensure
backwards compatibility, it is best to keep the original partition
layout.

If a larger, single firmware partition, is desired, one can still
install the XGS1210-12 firmware onto the switch.

In addition to the above, the XGS1010-12 lacks a physical reset button,
but all circuitry is in place to support this, including the whole in
the metal housing. As such, we keep the dts entry for the button, to
allow the user to solder on a button, or short the pins, if they so
desire.

As this switch does not feature any form of management or firmware
updateability, one must use the serial UART either through XYZ-modem
(untested) or tftp. The instructions to do so are identical as with the
XGS1250-12.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>