realtek: clock driver: simplify magic register values
For simplicity we use register magic values. We simply read/write full registers
but only some of the bits are relevant. So we might overwrite data that is
different depending on the vendor. For simplification we will handle constants
as follows:

- constants are right shifted and contain only relevant bits
- the code will left shift the values as needed
- the code will run a read-modify-write cycle
- As syncs are not needed for registers we can drop that part

Btw. SRAM can be access cached/uncached. When using parts of it make clear that
we access the cached part.

Signed-off-by: Markus Stockhausen <markus.stockhausen@gmx.de>