realtek: clock driver: provide crystal/oscillator clock
Until now the driver depends on an external DT oscillator definition.
For now we can not see any benefit of setting that in the device tree.
Lots of other Linux kernel drivers hardcode these values. Usually the
name xtal (crystal) is used there. Align that and register a fixed
frequency clock (named xtal_clk) in the driver.

Signed-off-by: Markus Stockhausen <markus.stockhausen@gmx.de>