realtek: Be a bit more careful when parsing the MAC addresses
We should check our pre-requisites better when setting the interface MAC
addresses.

For one, when we don't have an end mac address for the LAN ports,
we should set the Local Addressing bit for the LAN ports.

Secondly, if the label MAC and lan MAC are identical, increment by one
to ensure we don't use the same MAC address.

One Edge-case (which is handled) is that if the range is shorter then
the number of ports, when we encounter this, any following mac address
will use Local Addressing.

Further more, lets follow the original line of reasoning on MAC
assignments. The MAC address on the label, supplied via `ethaddr` should
be assigned to eth0 and the switch bridge, and 'mac_start - mac_end', if
available used for the individual lan ports. If those are not supplied,
use Link Local Addressing.

While here, make sure the new implementation is readable, shellcheck
compliant shell scripting, rather then cryptic, optimized for shortest
size.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>