base-files: Add MAC address validation helper
Rather then checking if a (mac address containing) variable is empty or
not, lets instead allow a user to validate a MAC address.

This function can be used in one of two ways. Either just called as a
function, where the return value (of grep) actually tells you if the MAC
address matches (0) or not (1). Alternatively, capture the output, the
MAC address matches (the MAC address is repeated) or not (empty).

Currently, this check only supports byte seperated mac addresses, but
could be extended to support word-dot seperated addresses as well. Since
this is however a format not really used by OpenWRT, lets keep it simple
for now.

The used regex, `^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]){2}$` can be put
into regexr.com or similar to get a nice textual explanation for
those curious.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>