realtek: Use SDK tx eye calibration tables
The SDK offers various calibration tables that they use as eye
calibration tables. These tables are not ideal, as they name the serdes
that they where intended for, with the function of the reference model,
never the less, it gives some meaning and structure to these magic
values.

These values should be set in the devicetree in the future, as they are
very much hardware dependent.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>