realtek: Calibrate TX SDS config for DAC cables
To configure the transmission serdes controller, we need to know first
if we are dealing with a DAC type, and secondly the length (longer DAC's
use the standard setting). This information is available on a DSA port's
PhyLink struct we can use, but have to pass a long.

This does not yet fix all DAC lengths, as the RX calibration is still
needed, but that is far from trivial.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>