realtek: Define CPU and port count properly
We currently have some magic values and defines laying around indicating
which port is the CPU port and how many ports we have in total. While
this is almost always identical, it is not what we wish to express. Just
because the CPU port is almost always the last port, does not mean this
is correct.

Instead, introduce a new define, PORT_CNT, which indicates the number of
ports a switch has. The CPU_PORT define then points to whatever port is
accurate, which in the case of the RTL931x is not the last one, as the
last 4 ports are combined to form one. Though this may need to become
configurable via the devicetree in the future, as it is possible not all
variations of the rtl931x series are like that.

With these defines we can replace some magic values making things a
little more clear.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>