realtek: Add missing family ID's
The realtek family chips often depend on the exact 16 bit family id. We
have a couple of missing IDs, so lets add those. While here, also sort
the existing entries.

Note, that the entire chip identification bit probably needs to be
redone in the future, as we care also about the lower 16bits of the ID.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>