realtek: eth: Use NIC defines
We have various defines that help with the NIC functionality. Lets
use those defines to be slightly less magical.

In the future, we probably want to switch out the GENMASK for something
more specific, although it seems that some of these masks related to a
bitfield of ports, which shouldn't even be here ...

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>