realtek: eth: Make NAPI polling thread safe
At the end of RX NAPI polling the counter and mask registers are
cleaned up. Although this might run in parallel there is no
synchronization and the register modifications are some wild mix.
So make use of the driver lock and only modify the bits that the
current thread owns.

Signed-off-by: Markus Stockhausen <markus.stockhausen@gmx.de>
Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>