realtek: Do not use sds reset for sds off
There is little point in using sds reset to turn the sds off, as reset
really just turns things off, and then on again, e.g. off wait off
again, doesn't serve any purpose.

If we need to turn an sds off twice, we should document and codify this
instead of hiding it in a reset.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>