realtek: Rename platform to just REALTEK_SOC
It makes little sense, to call our platform RTL83xx when in fact, it
also contains RTL93xx and in the future could contain RTL819x easily as
well. Lets do what we do for most platforms, and refer to the vendor
instead. Since realtek makes a lot of products (PHY's, USB peripherials,
Networking etc etc) based name it REALTEK_SOC to indicate what we mean
more accurately.

We did not choose to use `MACH_REALTEK_RTL` as is currently in upstream,
as that name makes even less sense, MACH not being appropriate, and `RTL`
being the well known abbreviation/prefix already, so in essence we are
only renaming from `MACH_REALTEK_RTL` to `REALTEK_SOC`.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>