realtek: clk: Cleanup defines
The current clock driver can get a little love and be more (future)
consistent.

The manual bitshifts where initially deemed needed because we didn't
have pre-processor support for assembly files, which technically is
true, however, we can still keep proper macro's but shadow the normal
ones in case we are assembly compiling stuff.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>