realtek: dts: Set proper SoC compatibles
Up until now, we did not actually use or care about the SoC compatible,
it was ignored and not used.

From now on however, we want to rely on the compatible to boot the right
configuration, and as such, we need to ensure the proper compatibles
have been set in the actual dts.

The dtsi can name the 'generic' compatible, which helps get users
started, but there is no point in having generic compatibles anywhere
else, as they will always be overwritten by the top-one.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>