realtek: eth: Transition to true fixed-link
We already put 'fixed-link' in our devicetree, yet we treat out CPU link
as if it was a normal 'auto-negotiable' link. While in theory this could
be true, in practice, the ethernet NIC doesn't have any capabilities to
configure these parameters. The MAC 'on the other end' however does, but
that is the DSA CPU port, and should be configured on that end.

While not fully known yet, it seems to be that the we are dealing with a
MAC <--> MAC link, and that the ethernet MAC just follows the Switch
MAC, either by using the same registers or something else. As such, lets
leave the configuration up to the DSA side of things, and instead just
do what the devicetree states for now.

This does make it also possible to far more simplify the phylinks ops,
as most are not needed anymore.

We can remove the `validate` op, as the built-in generic already did what
we where doing manually anyway.

We can remove the `mac_pcs_get_state` op, as this was never called to begin
with. The function only gets called if `legacy_pre_march2020` is set to
`true`, which in our case was unset and thus false, but even then, the
whole `phylink_mac_pcs_get_state` only ever gets called if
`MLO_AN_INBAND` is set, which is only set if we are not a `fixed-link`
which we are according to our devicetree and that is parsed as such by
`phylink_create`.

We can remove `mac_an_restart` op, as also this gets ignored if
`legacy_pre_march2020` is not set and thus has always been a NOOP. We
may have to force-off `an_enabled` or disable autonegotion advertisement
in the future however, depending on whether the MAC follows the NIC or
vice versa.

We must keep `mac_config`, `mac_link_down` and `mac_link_up` however as
they are mandatory, and not having them will cause NULL pointer exceptions.

We can however turn them to NOOPs as technically, we can't turn the link
on or off anyway, only the MAC on the switch port (CPU_PORT) end.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>