realtek: dts: Add i2c controllers to rtl930x main dts
The I2C master controller for the two I2C controllers is missing from
the devicetree. Lets add them so we can re-use this more.

The I2C driver still needs to be refactored to properly support
devicetree inclusion and the mux still needs to be made 'mandatory' (or
hidden, but still internally used), but that is out of scope for this
change.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>