realtek: dts: Reorganize busses on rtl930x dtsi
The realtek SoC is a strange beast, where we have two major ranges.
0x18000000 which is also known as the 'lexra peripherial bus'. These
peripherials seem to be mostly/all off-the-shelf peripherials. There is
also a the lexra 0 bus, which sits at 0x1b000000 and contains the
seemingly realtek in-house developed peripherals, such as the Ethernet
controller, I2C controllers, SPI controllers, RTL8231 LED and GPIO
external controller and the switch controller, with all its level 2 and
3 features.

Further more there is also a lexra bus 1, which seems to be unused and
the lexra bus 2, which contains the USB controller, ECC offload
controller, SPI-Nand controller and a few more.

Looking at the register layout however, it seems more accurate that the
lexra peripherial bus is where all 'control' registers are at the least,
as all the peripherials that are not on lexra bus 0, have all addresses
from before that. So for now, we 'pretend' all peripherials not on the
lexra bus 0, are part of the peripherial bus, and the rest is on lexra
bus 0.

As such, we should split our device tree to map to these busses with the
peripherials assigned underneath this structure. We no longer re-map our
ranges however, but instead opt to use the datasheet addressing, which
uses full-addresses, to make things slightly easier, and a little less
obfuscated.

While here, cleanup the touched area's a little bit, without affecting
functionality.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>