realtek: dts: Move internal rtl930x ethernet port dtsi
The Ethernet port is always connected to switch-port 28. We can keep
this in our main dtsi.

While here, fix the inconsistency between the Ethernet fixed link speed,
and the switch-port fixed link speed. These should be in sync. The speed
doesn't matter in the end anyway, as its only used for display purposes.
The Ethernet controller is not able to change speeds anyway.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>