realtek: eth: Add MDIO defines
We can demagic stuff with these defines. Note, that they are under the
more rtl930x header under dsa, as that's where most of the networking
related bits live right now.

In the future, a) the true MDIO related registers will end up in the
mdio specific regmaps in the mdio C file and b) the special MDIO
accelerated cases, which are switch related anyway, will end up in the
dsa bit anyhow.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>