realtek: dts: Rename spi0 to qspif0 on rtl93xx
The current 'spi' controller is actually not a regular SPI controller,
but a so called 'Quad SPI Flash' controller, or QSPIF. The main
difference is that they are intended to be used with SPI flash chips. In
such that they support dual or quad IO modes, as well as having support
for basic flash read commands, to memory-map the flash chips to do
In memory Execution (XIP, eXectute In Place).

While in theory one can use the QSPI controller for other normal things
as well, it's not realistic on these platforms, especially with our
usecase.

So lets rename the dts entries, so we can more easily add the actual SPI
controller this platform also has.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>