realtek: mdio: Move out of ethernet on rtl93xx
Traditionally, mdio controllers are part of, or sit within the Ethernet
controller. This is naturally where the MAC resides, which has this IP,
and needs to control the PHY.

With a switch, this is not true any more. The switch has multiple MAC's
all need to drive their own PHY's. As such, MDIO busses can happily sit
at the root of the bus, as is done with other switch chips as well.

To continue supporting older devicetree's (for now) that have their
nodes as ethernet children, we parse both. When upstreaming a proper
mdio driver, this would go away, and only the new, main, mdio node would
remain.

While here, drop the unused and unneeded regmap property.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>